Address second-pass review comments on classifier_context
- Change classifier_context type from Option<usize> to Option<u32>
  for consistent cross-platform config serialization
- Restructure window building to walk from the end when a bounded
  context is set, avoiding a full-history scan for small N values
- Fix debug log message: 'no user message' -> 'classifier input
  unavailable' since bypass can also happen for empty window or
  context=0, not just missing user messages
- Document 0 and 1 special-case semantics in configuration.md
  (0 = skip classification, 1 = last message only)

+++ src/config/profile.rs
@@ -178,7 +178,7 @@ pub struct ProfileConfig {
     /// context awareness (e.g. Home Assistant voice, where each utterance is
     /// mostly self-contained).
     #[serde(default)]
-    pub classifier_context: Option<usize>,
+    pub classifier_context: Option<u32>,
 
     /// Static response text returned by `reply` mode without calling any backend.
     ///
